+++ sense_hat/__init__.py
@@ -13,4 +13,4 @@ from .stick import (
     ACTION_HELD,
     )
 
-__version__ = '2.2.0'
+__version__ = '2.4.0'